utils/tools/otel_metadata_service/requirements.txt
# Dependencies for OTel PostgreSQL metadata fetcher service
# Core dependencies
httpx>=0.24.0
pyyaml>=6.0

# FastAPI web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

